+++ src/Commands/StopCommand.php
@@ -3,7 +3,6 @@
 namespace App\Commands;
 
 use App\Services\LogTimer;
-use App\Repositories\TaskRepository;
 use App\Services\Validators\StopValidator;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputInterface;
@@ -13,7 +12,6 @@ use Symfony\Component\Console\Output\OutputInterface;
  * Class StopCommand
  *
  * @author Mohamed Abdul-Fattah <csmohamed8@gmail.com>
- * @since  0.1.0
  */
 class StopCommand extends Command
 {
@@ -24,12 +22,14 @@ class StopCommand extends Command
 
     /**
      * StopCommand constructor.
+     *
+     * @param LogTimer $timer
      */
-    public function __construct()
+    public function __construct(LogTimer $timer)
     {
         parent::__construct();
 
-        $this->timer = new LogTimer(new TaskRepository);
+        $this->timer = $timer;
     }
 
     /**
@@ -50,23 +50,29 @@ class StopCommand extends Command
                  'd',
                  InputOption::VALUE_REQUIRED,
                  'Task log description'
+             )->addOption(
+                 'group',
+                 'g',
+                 InputOption::VALUE_OPTIONAL,
+                 'Tempo attributes group'
              );
     }
 
     /**
-     * @param  InputInterface $input
+     * @param  InputInterface  $input
      * @param  OutputInterface $output
-     * @return int|void
+     * @return int
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $desc      = $input->getOption('description');
         $time      = $input->getOption('time');
+        $group     = $input->getOption('group');
         $validator = new StopValidator($time);
         $validator->validate();
 
         $output->writeln('<comment>Stop logging...</comment>');
-        $this->timer->stop($time, $desc);
+        $this->timer->stop($time, $desc, $group);
         $output->writeln('<info>Log stopped successfully</info>');
 
         return self::EXIT_SUCCESS;