+++ src/Repositories/Contracts/ITaskRepository.php
@@ -8,29 +8,30 @@ use App\Entities\Task;
  * Interface ITaskRepository
  *
  * @author Mohamed Abdul-Fattah <csmohamed8@gmail.com>
- * @since  0.1.0
  */
 interface ITaskRepository
 {
     /**
      * Insert a running log with start time
      *
-     * @param  string $taskId
-     * @param  string $time
-     * @param  string $desc
+     * @param  string   $taskId
+     * @param  string   $time
+     * @param  string   $desc
+     * @param  int|null $groupId
      * @return void
      */
-    public function startLog($taskId, $time, $desc): void;
+    public function startLog($taskId, $time, $desc, $groupId = null): void;
 
     /**
      * Update running task with end time
      *
-     * @param  string $end
-     * @param  string $log
+     * @param  string      $end
+     * @param  string      $log
      * @param  string|null $desc
+     * @param  int|null    $groupId
      * @return void
      */
-    public function stopLog($end, $log, $desc = null): void;
+    public function stopLog($end, $log, $desc = null, $groupId = null): void;
 
     /**
      * Delete the current running task log