+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Dependency injector
+- `tempo:attributes` command to save extra attributes for logging
+- `tempo:list` command to list saved tempo attributes
+- Save tempo group on log start
+- Save tempo group on log stop
+- `tempo:sync` to sync logs to Jira via Tempo
+
 ## [0.3.1] - 2020-04-23
 ### Fixed
 - Fixed release checker
@@ -20,7 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.1] - 2020-03-09
 ### Fixed
-- Error on composer create-project because of overcommit with no git repo
+- Error on the composer create-project because of overcommit with no git repo
 
 ## [0.2.0] - 2020-03-08
 ### Added